Create entity::arrow module, move spawn logic there
