+++ server/src/packet_handlers/digging.rs
@@ -4,23 +4,17 @@
 //! for actions mostly unrelated to digging including eating, shooting bows,
 //! swapping items out to the offhand, and dropping items.
 
-use crate::entity;
-use crate::entity::{ComponentSerializer, EntityId, SpawnPacketCreator, Velocity};
+use crate::entity::arrow;
 use crate::game::Game;
 use crate::p_inventory::EntityInventory;
 use crate::packet_buffer::PacketBuffers;
-use crate::physics::{charge_from_ticks_held, compute_projectile_velocity, PhysicsBuilder};
+use crate::physics::{charge_from_ticks_held, compute_projectile_velocity};
 use crate::player::{ItemTimedUse, PLAYER_EYE_HEIGHT};
-use crate::util::{degrees_to_stops, protocol_velocity};
-use feather_core::entity::{ArrowEntityData, BaseEntityData, EntityData};
 use feather_core::inventory::{SlotIndex, SLOT_HOTBAR_OFFSET, SLOT_OFFHAND};
-use feather_core::network::packet::implementation::{
-    PlayerDigging, PlayerDiggingStatus, SpawnObject,
-};
-use feather_core::{Block, Gamemode, Item, ItemStack, Packet, Position, Vec3d};
-use fecs::{Entity, EntityRef, World};
+use feather_core::network::packet::implementation::{PlayerDigging, PlayerDiggingStatus};
+use feather_core::{Block, Gamemode, Item, ItemStack, Position};
+use fecs::{Entity, World};
 use std::sync::Arc;
-use uuid::Uuid;
 
 /// System responsible for polling for PlayerDigging
 /// packets and writing the corresponding events.
@@ -219,7 +213,7 @@ fn handle_shoot_bow(game: &mut Game, world: &mut World, player: Entity) {
     }
 
     let charge_force = charge_from_ticks_held(time_held as u32);
-    debug!("Held for {} ticks. Force of {}", time_held, charge_force);
+    trace!("Held for {} ticks. Force of {}", time_held, charge_force);
 
     let init_position = *world.get::<Position>(player) + glm::vec3(0.0, PLAYER_EYE_HEIGHT, 0.0);
 
@@ -231,7 +225,7 @@ fn handle_shoot_bow(game: &mut Game, world: &mut World, player: Entity) {
         0.0,
         &mut *game.rng(),
     );
-    debug!(
+    trace!(
         "Computed exit velocity: {}. Velocity is norm {}",
         arrow_velocity,
         arrow_velocity.norm()
@@ -241,23 +235,8 @@ fn handle_shoot_bow(game: &mut Game, world: &mut World, player: Entity) {
 
     world.remove::<ItemTimedUse>(player).unwrap();
 
-    debug!("Preparing to spawn entity");
-
-    let entity = entity::base(init_position)
-        .with(Velocity(arrow_velocity))
-        .with(SpawnPacketCreator(&create_spawn_packet))
-        .with(ComponentSerializer(&serialize))
-        .with(
-            PhysicsBuilder::new()
-                .bbox(0.5, 0.5, 0.5)
-                .gravity(-0.05)
-                .slip_multiplier(0.0)
-                .drag(0.99)
-                .build(),
-        )
-        .build()
-        .spawn_in(world);
-    game.on_entity_spawn(world, entity);
+    trace!("Spawning arrow entity.");
+    arrow::spawn_arrow(game, world, arrow_velocity, init_position);
 }
 
 fn find_arrow(inventory: &EntityInventory) -> Option<(SlotIndex, ItemStack)> {
@@ -292,38 +271,4 @@ fn is_arrow_item(item: Item) -> bool {
         Item::Arrow | Item::SpectralArrow | Item::TippedArrow => true,
         _ => false,
     }
-}
-
-fn create_spawn_packet(accessor: &EntityRef) -> Box<dyn Packet> {
-    let position = *accessor.get::<Position>();
-    let velocity = *accessor.get::<Velocity>();
-    let entity_id = accessor.get::<EntityId>().0;
-
-    let (velocity_x, velocity_y, velocity_z) = protocol_velocity(velocity.0);
-
-    let packet = SpawnObject {
-        entity_id,
-        object_uuid: Uuid::new_v4(),
-        ty: 60, // Type 60 for arrow projectile
-        x: position.x,
-        y: position.y,
-        z: position.z,
-        pitch: degrees_to_stops(position.pitch),
-        yaw: degrees_to_stops(position.yaw),
-        data: entity_id + 1,
-        velocity_x,
-        velocity_y,
-        velocity_z,
-    };
-
-    Box::new(packet)
-}
-
-fn serialize(_game: &Game, accessor: &EntityRef) -> EntityData {
-    let vel = accessor.get::<Velocity>();
-
-    EntityData::Arrow(ArrowEntityData {
-        entity: BaseEntityData::new(*accessor.get::<Position>(), Vec3d::new(vel.x, vel.y, vel.z)),
-        critical: 0, // TODO
-    })
 }
